requirements.txt
# PettingZoo[classic,butterfly]>=1.22.3
# Pillow>=9.4.0
# ray[rllib]>=2.2.0
# SuperSuit>=3.2.0
# torch>=1.13.1
# tensorflow-probability>=0.19.0
# gymnasium==0.26.1

# PettingZoo[classic,butterfly]>=1.22.3
# Pillow>=9.4.0
# ray[rllib]>=2.5.1
# SuperSuit<=3.9.0
# torch>=1.13.1
# tensorflow-probability>=0.19.0

PettingZoo[classic, butterfly]==1.22.3
Pillow>=9.4.0
ray[rllib]>=2.6.2
SuperSuit==3.8.0
torch>=1.13.1
tensorflow-probability>=0.19.0
optuna>=3.3.0
GPUtil>=1.4.0
ipywidgets>=8.1.0

